Revert some of utility class related changes
