+++ platforms/core-configuration/declarative-dsl-core/src/test/kotlin/org/gradle/internal/declarativedsl/dom/mutation/MutationAsTextRunnerTest.kt
@@ -32,6 +32,7 @@ import org.gradle.internal.declarativedsl.schemaUtils.findPropertyFor
 import org.gradle.internal.declarativedsl.schemaUtils.findTypeFor
 import org.gradle.internal.declarativedsl.schemaUtils.functionFor
 import org.gradle.internal.declarativedsl.schemaUtils.propertyFor
+import org.gradle.internal.declarativedsl.schemaUtils.typeFor
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.Test
