+++ platforms/core-configuration/declarative-dsl-provider/src/test/kotlin/org/gradle/internal/declarativedsl/ContainersSchemaComponentTest.kt
@@ -30,14 +30,14 @@ import org.gradle.declarative.dsl.schema.FunctionSemantics
 import org.gradle.internal.declarativedsl.analysis.SchemaTypeRefContext
 import org.gradle.internal.declarativedsl.analysis.analyzeEverything
 import org.gradle.internal.declarativedsl.common.gradleDslGeneralSchema
-import org.gradle.internal.declarativedsl.dom.mutation.anyType
-import org.gradle.internal.declarativedsl.dom.mutation.typeFor
+import org.gradle.internal.declarativedsl.schemaUtils.typeFor
 import org.gradle.internal.declarativedsl.evaluationSchema.SimpleInterpretationSequenceStepWithConversion
 import org.gradle.internal.declarativedsl.evaluationSchema.buildEvaluationAndConversionSchema
 import org.gradle.internal.declarativedsl.evaluator.conversion.AnalysisAndConversionStepRunner
 import org.gradle.internal.declarativedsl.evaluator.conversion.ConversionStepContext
 import org.gradle.internal.declarativedsl.evaluator.runner.AnalysisStepContext
 import org.gradle.internal.declarativedsl.evaluator.runner.AnalysisStepRunner
+import org.gradle.internal.declarativedsl.schemaUtils.findType
 import org.gradle.internal.declarativedsl.schemaUtils.hasFunctionNamed
 import org.gradle.internal.declarativedsl.schemaUtils.singleFunctionNamed
 import org.gradle.internal.reflect.Instantiator
@@ -56,7 +56,7 @@ class ContainersSchemaComponentTest {
     @Test
     fun `uses element type names for factory function names`() {
         listOf("one", "two", "customFactoryName").forEach { name ->
-            assertTrue(schema.analysisSchema.anyType { type -> type.hasFunctionNamed(name) })
+            assertTrue(schema.analysisSchema.findType { type: DataClass -> type.hasFunctionNamed(name) } != null)
         }
     }
